Add ClonableMixin
This is to allow a cloning feature of a model. It is possible to either clone and make the clone a new instance which is already persisted in the db or not. Everything depends on how you call the clone method. Example
```python
myModel = Article.objects.create(name='a test name')
print(myModel.pk)  # 1
article2 = myModel.clone()
print(article2.pk)  # None
article3 = myModel.clone(True) 
print(article3.pk)  # 2
```